Moved the run call out of the src folder
