+++ action.yml
@@ -25,4 +25,4 @@ outputs:
 
 runs:
   using: node12
-  main: dist/index.js
+  main: index.js
